Fixed issue (#18) on ReadInt64 method of StreamRW class.

+++ src/StreamRW.cs
@@ -50,7 +50,7 @@ namespace OpenMcdf
             this.stream.Read(buffer, 0, 8);
             uint ls = (uint)(buffer[0] | (buffer[1] << 8) | (buffer[2] << 16) | (buffer[3] << 24));
             uint ms = (uint)((buffer[4]) | (buffer[5] << 8) | (buffer[6] << 16) | (buffer[7] << 24));
-            return (long)((ms << 32) | ls);
+            return (long)(((ulong)ms << 32) | ls);
         }
 
         public ulong ReadUInt64()
